+++ themes/default/views/reports/ReportOfStudentsBenefitingFromTheBF.php
@@ -3,7 +3,7 @@
 /* @var $report mixed */
 $baseUrl = Yii::app()->baseUrl;
 $cs = Yii::app()->getClientScript();
-$cs->registerScriptFile($baseUrl . '/js/reports/BFReport/_initialization.js', CClientScript::POS_END);
+$cs->registerScriptFile($baseUrl . '/js/reports/ReportOfStudentsBenefitingFromTheBF/_initialization.js', CClientScript::POS_END);
 
 $this->setPageTitle('TAG - ' . Yii::t('default', 'Reports'));
 ?>
@@ -20,20 +20,20 @@ $this->setPageTitle('TAG - ' . Yii::t('default', 'Reports'));
     </div>
     <table class="table table-bordered table-striped" aria-labelledby="cns students">
         <thead>
-            <?php echo $allSchools ? "<th scope='col'>ESCOLA</th>" : ""?>
+            <?php echo $allSchools ? "<th scope='col'>ESCOLA</th>" : "<th scope='col'>TURMA</th>"?>
             <th>NOME</th>
             <th>D.N</th>
-            <th>CNS</th>
+            <th>NIS</th>
             <th>RESPONSÁVEL</th>
             <th>TEL. RESPONSÁVEL</th>
         </thead>
         <tbody>
             <?php foreach($report as $r) { ?>
                 <tr>
-                    <?php echo $allSchools ? "<td>".$r['school_name']."</td>" : ""?>
+                    <?php echo $allSchools ? "<td>".$r['school_name']."</td>" : "<td>".$r['classroom_name']."</td>"?>
                     <td><?= $r['name'] ?></td>
-                    <td><?= TagUtils::convertDateFormat($r['birthday']) ?></td>
-                    <td><?= $r['cns'] ?></td>
+                    <td><?= TagUtils::convertDateFormat($r['birthday'])?></td>
+                    <td><?= $r['nis'] ?></td>
                     <td><?= $r['responsable_name'] ?></td>
                     <td><?= $r['responsable_telephone'] ?></td>
                 </tr>